Continuing work on LED driver code
Still not working.... I broke the code up into header files. Which makes
it much nicer to read, however, OO was the only way to go and I'm afraid
that it generates longer code. Will need to investigate this. Can light
an arbitrary selection of LEDs on, don't know why some are off, can't
turn them off. Much work yet to be done.